requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
flask>=3.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
playwright>=1.40.0
